requirements.txt
accelerate==1.0.1
bitsandbytes==0.44.1
colorama==0.4.6
datasets==3.0.0
google-generativeai==0.8.3
guidance==0.2.0rc1
jsonschema==4.23.0
matplotlib==3.9.2
numpy==1.26.4
omegaconf==2.3.0
openai==1.46.0
outlines==0.1.8
pandas==2.2.2
psutil==6.0.0
pydantic==2.9.2
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
stopit==1.1.2
termcolor==2.4.0
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
xgrammar==0.1.6
